+++ fj-core/src/test/java/test/org/fugerit/java/core/db/daogein/TestSQLTypeConverter.java
@@ -7,8 +7,11 @@ import org.junit.Test;
 
 import java.sql.Date;
 import java.sql.Timestamp;
+import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.temporal.TemporalField;
+import java.util.Calendar;
 
 @Slf4j
 public class TestSQLTypeConverter {
@@ -20,6 +23,11 @@ public class TestSQLTypeConverter {
         LocalDate d = SQLTypeConverter.utilDateToLocalDate(Date.valueOf( TEST_YEAR+"-05-05" ));
         log.info( "local date : {}", d );
         Assert.assertEquals( TEST_YEAR, d.getYear() );
+        Date t = SQLTypeConverter.localDateToSqlDate( d );
+        log.info( "sql date : {}", t );
+        Calendar c = Calendar.getInstance();
+        c.setTime(t);
+        Assert.assertEquals( TEST_YEAR, c.get( Calendar.YEAR ) );
     }
 
     @Test
@@ -27,6 +35,11 @@ public class TestSQLTypeConverter {
         LocalDateTime d = SQLTypeConverter.utilDateToLocalDateTime(Timestamp.valueOf( TEST_YEAR+"-05-05 11:30:00.000" ));
         log.info( "local date time : {}", d );
         Assert.assertEquals( TEST_YEAR, d.getYear() );
+        Timestamp t = SQLTypeConverter.localDateTimeToTimestamp( d );
+        log.info( "timestamp : {}", t );
+        Calendar c = Calendar.getInstance();
+        c.setTime(t);
+        Assert.assertEquals( TEST_YEAR, c.get( Calendar.YEAR ) );
     }
 
 }